Añadiendo modelos para funcion de POST de cada novedad poscontractual #4
