+++ controllers/novedades.go
@@ -3,7 +3,6 @@ package controllers
 import (
 	"encoding/json"
 	"fmt"
-	"strconv"
 
 	"github.com/astaxie/beego"
 	"github.com/udistrital/novedades_mid/models"
@@ -139,7 +138,7 @@ func RegistrarNovedad(novedad map[string]interface{}) (status interface{}, outpu
 	case "59d7965e867ee188e42d8c72":
 		//suspensión
 		fmt.Println("Novedad de suspensión")
-		NovedadPoscontractualPost = ConstruirNovedadSuspension(registroNovedadPost)
+		NovedadPoscontractualPost = models.ConstruirNovedadSuspension(registroNovedadPost)
 		fmt.Println(NovedadPoscontractualPost)
 	case "59d79683867ee188e42d8c97":
 		//cesión
@@ -147,26 +146,27 @@ func RegistrarNovedad(novedad map[string]interface{}) (status interface{}, outpu
 	case "59d796ac867ee188e42d8cbf":
 		//reinicio
 		fmt.Println("Novedad de reinicio")
-	case "59d797aa867ee188e42d8db6":
-		//liquidación
-		fmt.Println("Novedad de liquidación")
+		NovedadPoscontractualPost = models.ConstruirNovedadReinicio(registroNovedadPost)
+		fmt.Println(NovedadPoscontractualPost)
 	case "59d79809867ee188e42d8e0d":
 		//terminacion anticipada
 		fmt.Println("Novedad de terminación anticipada")
+		NovedadPoscontractualPost = models.ConstruirNovedadTAnticipada(registroNovedadPost)
+		fmt.Println(NovedadPoscontractualPost)
 	case "59d7985e867ee188e42d8e64":
 		//adición
 		fmt.Println("Novedad de adición")
-		NovedadPoscontractualPost = ConstruirNovedadAdicionPost(registroNovedadPost)
+		NovedadPoscontractualPost = models.ConstruirNovedadAdicionPost(registroNovedadPost)
 		fmt.Println(NovedadPoscontractualPost)
 	case "59d79894867ee188e42d8e9b":
 		//prórroga
 		fmt.Println("Novedad de prorroga")
-		NovedadPoscontractualPost = ConstruirNovedadProrrogaPost(registroNovedadPost)
+		NovedadPoscontractualPost = models.ConstruirNovedadProrrogaPost(registroNovedadPost)
 		fmt.Println(NovedadPoscontractualPost)
 	case "59d79904867ee188e42d8f02":
 		//adicion/prorroga
 		fmt.Println("Novedad de adicion/prorroga")
-		NovedadPoscontractualPost = ConstruirNovedadAdProrrogaPost(registroNovedadPost)
+		NovedadPoscontractualPost = models.ConstruirNovedadAdProrrogaPost(registroNovedadPost)
 		fmt.Println(NovedadPoscontractualPost)
 	}
 
@@ -203,464 +203,4 @@ func ConsultarNovedadID(idNovedad string) (novedad map[string]interface{}, outpu
 	errnovedad := request.GetJson("http://"+beego.AppConfig.String("NovedadesCrudService")+"/v1/novedades_poscontractuales/?query=ContratoId:"+novedadid, &getnovedad)
 
 	return nil, errnovedad
-}
-
-func ConstruirNovedadAdicionPost(novedad map[string]interface{}) (novedadformatted map[string]interface{}) {
-	NovedadAdicion := make(map[string]interface{})
-	NovedadAdicion = novedad
-
-	NovedadAdicionPost := make(map[string]interface{})
-	contratoid, _ := strconv.ParseInt(NovedadAdicion["contrato"].(string), 10, 32)
-	numerocdpid, _ := strconv.ParseInt(NovedadAdicion["numerocdp"].(string), 10, 32)
-	numerosolicitudentero := NovedadAdicion["numerosolicitud"].(float64)
-	numerosolicitud := strconv.FormatFloat(numerosolicitudentero, 'f', -1, 64)
-	vigencia, _ := strconv.ParseInt(NovedadAdicion["vigencia"].(string), 10, 32)
-	vigenciacdp, _ := strconv.ParseInt(NovedadAdicion["vigencia"].(string), 10, 32)
-
-	fmt.Println(NovedadAdicion["contrato"], NovedadAdicion["numerocdp"], NovedadAdicion["numerosolicitud"], NovedadAdicion["vigencia"], NovedadAdicion["vigencia"])
-	fmt.Println("\n", contratoid, numerocdpid, numerosolicitud, vigencia, vigenciacdp, "\n")
-
-	NovedadAdicionPost["NovedadPoscontractual"] = map[string]interface{}{
-		"Aclaracion":        nil,
-		"Activo":            true,
-		"ContratoId":        contratoid,
-		"FechaCreacion":     nil,
-		"FechaModificacion": nil,
-		"Id":                0,
-		"Motivo":            NovedadAdicion["motivo"],
-		"NumeroCdpId":       numerocdpid,
-		"NumeroSolicitud":   numerosolicitud,
-		"Observacion":       nil,
-		"TipoNovedad":       6,
-		"Vigencia":          vigencia,
-		"VigenciaCdp":       vigenciacdp,
-	}
-
-	fechas := make([]map[string]interface{}, 0)
-
-	fechas = append(fechas, map[string]interface{}{
-		"Activo":            true,
-		"Fecha":             NovedadAdicion["fechasolicitud"],
-		"FechaCreacion":     nil,
-		"FechaModificacion": nil,
-		"Id":                0,
-		"IdNovedadesPoscontractuales": map[string]interface{}{
-			"Id": nil,
-		},
-		"IdTipoFecha": map[string]interface{}{
-			"Id": 7,
-		},
-	})
-	fechas = append(fechas, map[string]interface{}{
-		"Activo":            true,
-		"Fecha":             NovedadAdicion["fechaadicion"],
-		"FechaCreacion":     nil,
-		"FechaModificacion": nil,
-		"Id":                0,
-		"IdNovedadesPoscontractuales": map[string]interface{}{
-			"Id": nil,
-		},
-		"IdTipoFecha": map[string]interface{}{
-			"Id": 1,
-		},
-	})
-
-	NovedadAdicionPost["Fechas"] = fechas
-
-	propiedades := make([]map[string]interface{}, 0)
-	propiedades = append(propiedades, map[string]interface{}{
-		"Activo":            true,
-		"FechaCreacion":     nil,
-		"FechaModificacion": nil,
-		"Id":                0,
-		"IdNovedadesPoscontractuales": map[string]interface{}{
-			"Id": nil,
-		},
-		"IdTipoPropiedad": map[string]interface{}{
-			"Id": 6,
-		},
-		"propiedad": NovedadAdicion["valoradicion"],
-	})
-
-	propiedades = append(propiedades, map[string]interface{}{
-		"Activo":            true,
-		"FechaCreacion":     nil,
-		"FechaModificacion": nil,
-		"Id":                0,
-		"IdNovedadesPoscontractuales": map[string]interface{}{
-			"Id": nil,
-		},
-		"IdTipoPropiedad": map[string]interface{}{
-			"Id": 2,
-		},
-		"propiedad": NovedadAdicion["cesionario"],
-	})
-
-	NovedadAdicionPost["Propiedad"] = propiedades
-
-	fmt.Println(NovedadAdicionPost)
-
-	return NovedadAdicionPost
-}
-
-func ConstruirNovedadProrrogaPost(novedad map[string]interface{}) (novedadformatted map[string]interface{}) {
-	NovedadProrroga := make(map[string]interface{})
-	NovedadProrroga = novedad
-
-	NovedadProrrogaPost := make(map[string]interface{})
-	contratoid, _ := strconv.ParseInt(NovedadProrroga["contrato"].(string), 10, 32)
-	numerocdpid, _ := strconv.ParseInt(NovedadProrroga["numerocdp"].(string), 10, 32)
-	numerosolicitudentero := NovedadProrroga["numerosolicitud"].(float64)
-	numerosolicitud := strconv.FormatFloat(numerosolicitudentero, 'f', -1, 64)
-	vigencia, _ := strconv.ParseInt(NovedadProrroga["vigencia"].(string), 10, 32)
-	vigenciacdp, _ := strconv.ParseInt(NovedadProrroga["vigencia"].(string), 10, 32)
-
-	NovedadProrrogaPost["NovedadPoscontractual"] = map[string]interface{}{
-		"Aclaracion":        nil,
-		"Activo":            true,
-		"ContratoId":        contratoid,
-		"FechaCreacion":     nil,
-		"FechaModificacion": nil,
-		"Id":                0,
-		"Motivo":            NovedadProrroga["motivo"],
-		"NumeroCdpId":       numerocdpid,
-		"NumeroSolicitud":   numerosolicitud,
-		"Observacion":       nil,
-		"TipoNovedad":       7,
-		"Vigencia":          vigencia,
-		"VigenciaCdp":       vigenciacdp,
-	}
-
-	fechas := make([]map[string]interface{}, 0)
-
-	fechas = append(fechas, map[string]interface{}{
-		"Activo":            true,
-		"Fecha":             NovedadProrroga["fechasolicitud"],
-		"FechaCreacion":     nil,
-		"FechaModificacion": nil,
-		"Id":                0,
-		"IdNovedadesPoscontractuales": map[string]interface{}{
-			"Id": nil,
-		},
-		"IdTipoFecha": map[string]interface{}{
-			"Id": 7,
-		},
-	})
-	fechas = append(fechas, map[string]interface{}{
-		"Activo":            true,
-		"Fecha":             "2019-10-08T15:43:51.710Z",
-		"FechaCreacion":     nil,
-		"FechaModificacion": nil,
-		"Id":                0,
-		"IdNovedadesPoscontractuales": map[string]interface{}{
-			"Id": nil,
-		},
-		"IdTipoFecha": map[string]interface{}{
-			"Id": 1,
-		},
-	})
-	fechas = append(fechas, map[string]interface{}{
-		"Activo":            true,
-		"Fecha":             NovedadProrroga["fechaprorroga"],
-		"FechaCreacion":     nil,
-		"FechaModificacion": nil,
-		"Id":                0,
-		"IdNovedadesPoscontractuales": map[string]interface{}{
-			"Id": nil,
-		},
-		"IdTipoFecha": map[string]interface{}{
-			"Id": 4,
-		},
-	})
-
-	NovedadProrrogaPost["Fechas"] = fechas
-
-	propiedades := make([]map[string]interface{}, 0)
-	propiedades = append(propiedades, map[string]interface{}{
-		"Activo":            true,
-		"FechaCreacion":     nil,
-		"FechaModificacion": nil,
-		"Id":                0,
-		"IdNovedadesPoscontractuales": map[string]interface{}{
-			"Id": nil,
-		},
-		"IdTipoPropiedad": map[string]interface{}{
-			"Id": 6,
-		},
-		"propiedad": NovedadProrroga["valoradicion"],
-	})
-
-	propiedades = append(propiedades, map[string]interface{}{
-		"Activo":            true,
-		"FechaCreacion":     nil,
-		"FechaModificacion": nil,
-		"Id":                0,
-		"IdNovedadesPoscontractuales": map[string]interface{}{
-			"Id": nil,
-		},
-		"IdTipoPropiedad": map[string]interface{}{
-			"Id": 5,
-		},
-		"propiedad": NovedadProrroga["tiempoprorroga"],
-	})
-
-	propiedades = append(propiedades, map[string]interface{}{
-		"Activo":            true,
-		"FechaCreacion":     nil,
-		"FechaModificacion": nil,
-		"Id":                0,
-		"IdNovedadesPoscontractuales": map[string]interface{}{
-			"Id": nil,
-		},
-		"IdTipoPropiedad": map[string]interface{}{
-			"Id": 2,
-		},
-		"propiedad": NovedadProrroga["cesionario"],
-	})
-
-	NovedadProrrogaPost["Propiedad"] = propiedades
-
-	fmt.Println(NovedadProrrogaPost)
-
-	return NovedadProrrogaPost
-}
-func ConstruirNovedadAdProrrogaPost(novedad map[string]interface{}) (novedadformatted map[string]interface{}) {
-	NovedadAdProrroga := make(map[string]interface{})
-	NovedadAdProrroga = novedad
-
-	NovedadAdProrrogaPost := make(map[string]interface{})
-	contratoid, _ := strconv.ParseInt(NovedadAdProrroga["contrato"].(string), 10, 32)
-	numerocdpid, _ := strconv.ParseInt(NovedadAdProrroga["numerocdp"].(string), 10, 32)
-	numerosolicitudentero := NovedadAdProrroga["numerosolicitud"].(float64)
-	numerosolicitud := strconv.FormatFloat(numerosolicitudentero, 'f', -1, 64)
-	vigencia, _ := strconv.ParseInt(NovedadAdProrroga["vigencia"].(string), 10, 32)
-	vigenciacdp, _ := strconv.ParseInt(NovedadAdProrroga["vigencia"].(string), 10, 32)
-
-	NovedadAdProrrogaPost["NovedadPoscontractual"] = map[string]interface{}{
-		"Aclaracion":        nil,
-		"Activo":            true,
-		"ContratoId":        contratoid,
-		"FechaCreacion":     nil,
-		"FechaModificacion": nil,
-		"Id":                0,
-		"Motivo":            NovedadAdProrroga["motivo"],
-		"NumeroCdpId":       numerocdpid,
-		"NumeroSolicitud":   numerosolicitud,
-		"Observacion":       nil,
-		"TipoNovedad":       8,
-		"Vigencia":          vigencia,
-		"VigenciaCdp":       vigenciacdp,
-	}
-
-	fechas := make([]map[string]interface{}, 0)
-
-	fechas = append(fechas, map[string]interface{}{
-		"Activo":            true,
-		"Fecha":             NovedadAdProrroga["fechasolicitud"],
-		"FechaCreacion":     nil,
-		"FechaModificacion": nil,
-		"Id":                0,
-		"IdNovedadesPoscontractuales": map[string]interface{}{
-			"Id": nil,
-		},
-		"IdTipoFecha": map[string]interface{}{
-			"Id": 7,
-		},
-	})
-	fechas = append(fechas, map[string]interface{}{
-		"Activo":            true,
-		"Fecha":             NovedadAdProrroga["fechaadicion"],
-		"FechaCreacion":     nil,
-		"FechaModificacion": nil,
-		"Id":                0,
-		"IdNovedadesPoscontractuales": map[string]interface{}{
-			"Id": nil,
-		},
-		"IdTipoFecha": map[string]interface{}{
-			"Id": 1,
-		},
-	})
-	fechas = append(fechas, map[string]interface{}{
-		"Activo":            true,
-		"Fecha":             NovedadAdProrroga["fechaprorroga"],
-		"FechaCreacion":     nil,
-		"FechaModificacion": nil,
-		"Id":                0,
-		"IdNovedadesPoscontractuales": map[string]interface{}{
-			"Id": nil,
-		},
-		"IdTipoFecha": map[string]interface{}{
-			"Id": 4,
-		},
-	})
-
-	NovedadAdProrrogaPost["Fechas"] = fechas
-
-	propiedades := make([]map[string]interface{}, 0)
-	propiedades = append(propiedades, map[string]interface{}{
-		"Activo":            true,
-		"FechaCreacion":     nil,
-		"FechaModificacion": nil,
-		"Id":                0,
-		"IdNovedadesPoscontractuales": map[string]interface{}{
-			"Id": nil,
-		},
-		"IdTipoPropiedad": map[string]interface{}{
-			"Id": 6,
-		},
-		"propiedad": NovedadAdProrroga["valoradicion"],
-	})
-
-	propiedades = append(propiedades, map[string]interface{}{
-		"Activo":            true,
-		"FechaCreacion":     nil,
-		"FechaModificacion": nil,
-		"Id":                0,
-		"IdNovedadesPoscontractuales": map[string]interface{}{
-			"Id": nil,
-		},
-		"IdTipoPropiedad": map[string]interface{}{
-			"Id": 5,
-		},
-		"propiedad": NovedadAdProrroga["tiempoprorroga"],
-	})
-
-	propiedades = append(propiedades, map[string]interface{}{
-		"Activo":            true,
-		"FechaCreacion":     nil,
-		"FechaModificacion": nil,
-		"Id":                0,
-		"IdNovedadesPoscontractuales": map[string]interface{}{
-			"Id": nil,
-		},
-		"IdTipoPropiedad": map[string]interface{}{
-			"Id": 2,
-		},
-		"propiedad": NovedadAdProrroga["cesionario"],
-	})
-
-	NovedadAdProrrogaPost["Propiedad"] = propiedades
-
-	fmt.Println(NovedadAdProrrogaPost)
-
-	return NovedadAdProrrogaPost
-}
-func ConstruirNovedadSuspension(novedad map[string]interface{}) (novedadformatted map[string]interface{}) {
-	NovedadSuspension := make(map[string]interface{})
-	NovedadSuspension = novedad
-
-	NovedadSuspensionPost := make(map[string]interface{})
-	contratoid, _ := strconv.ParseInt(NovedadSuspension["contrato"].(string), 10, 32)
-	//numerocdpid, _ := strconv.ParseInt(NovedadSuspension["numerocdp"].(string), 10, 32)
-	numerosolicitudentero := NovedadSuspension["numerosolicitud"].(float64)
-	numerosolicitud := strconv.FormatFloat(numerosolicitudentero, 'f', -1, 64)
-	vigencia, _ := strconv.ParseInt(NovedadSuspension["vigencia"].(string), 10, 32)
-	vigenciacdp, _ := strconv.ParseInt(NovedadSuspension["vigencia"].(string), 10, 32)
-
-	NovedadSuspensionPost["NovedadPoscontractual"] = map[string]interface{}{
-		"Aclaracion":        nil,
-		"Activo":            true,
-		"ContratoId":        contratoid,
-		"FechaCreacion":     nil,
-		"FechaModificacion": nil,
-		"Id":                0,
-		"Motivo":            NovedadSuspension["motivo"],
-		"NumeroCdpId":       0,
-		"NumeroSolicitud":   numerosolicitud,
-		"Observacion":       nil,
-		"TipoNovedad":       1,
-		"Vigencia":          vigencia,
-		"VigenciaCdp":       vigenciacdp,
-	}
-
-	fechas := make([]map[string]interface{}, 0)
-
-	fechas = append(fechas, map[string]interface{}{
-		"Activo":            true,
-		"Fecha":             NovedadSuspension["fecharegistro"],
-		"FechaCreacion":     nil,
-		"FechaModificacion": nil,
-		"Id":                0,
-		"IdNovedadesPoscontractuales": map[string]interface{}{
-			"Id": nil,
-		},
-		"IdTipoFecha": map[string]interface{}{
-			"Id": 5,
-		},
-	})
-	fechas = append(fechas, map[string]interface{}{
-		"Activo":            true,
-		"Fecha":             NovedadSuspension["fechasolicitud"],
-		"FechaCreacion":     nil,
-		"FechaModificacion": nil,
-		"Id":                0,
-		"IdNovedadesPoscontractuales": map[string]interface{}{
-			"Id": nil,
-		},
-		"IdTipoFecha": map[string]interface{}{
-			"Id": 7,
-		},
-	})
-	fechas = append(fechas, map[string]interface{}{
-		"Activo":            true,
-		"Fecha":             NovedadSuspension["fechasuspension"],
-		"FechaCreacion":     nil,
-		"FechaModificacion": nil,
-		"Id":                0,
-		"IdNovedadesPoscontractuales": map[string]interface{}{
-			"Id": nil,
-		},
-		"IdTipoFecha": map[string]interface{}{
-			"Id": 8,
-		},
-	})
-	fechas = append(fechas, map[string]interface{}{
-		"Activo":            true,
-		"Fecha":             NovedadSuspension["fechareinicio"],
-		"FechaCreacion":     nil,
-		"FechaModificacion": nil,
-		"Id":                0,
-		"IdNovedadesPoscontractuales": map[string]interface{}{
-			"Id": nil,
-		},
-		"IdTipoFecha": map[string]interface{}{
-			"Id": 6,
-		},
-	})
-
-	NovedadSuspensionPost["Fechas"] = fechas
-
-	propiedades := make([]map[string]interface{}, 0)
-	propiedades = append(propiedades, map[string]interface{}{
-		"Activo":            true,
-		"FechaCreacion":     nil,
-		"FechaModificacion": nil,
-		"Id":                0,
-		"IdNovedadesPoscontractuales": map[string]interface{}{
-			"Id": nil,
-		},
-		"IdTipoPropiedad": map[string]interface{}{
-			"Id": 3,
-		},
-		"propiedad": NovedadSuspension["periodosuspension"],
-	})
-	propiedades = append(propiedades, map[string]interface{}{
-		"Activo":            true,
-		"FechaCreacion":     nil,
-		"FechaModificacion": nil,
-		"Id":                0,
-		"IdNovedadesPoscontractuales": map[string]interface{}{
-			"Id": nil,
-		},
-		"IdTipoPropiedad": map[string]interface{}{
-			"Id": 2,
-		},
-		"propiedad": NovedadSuspension["cesionario"],
-	})
-
-	NovedadSuspensionPost["Propiedad"] = propiedades
-
-	fmt.Println(NovedadSuspensionPost)
-
-	return NovedadSuspensionPost
 }
