+++ models/adicion.go
@@ -1,0 +1,104 @@
+package models
+
+import (
+	"fmt"
+	"strconv"
+)
+
+func ConstruirNovedadAdicionPost(novedad map[string]interface{}) (novedadformatted map[string]interface{}) {
+	NovedadAdicion := make(map[string]interface{})
+	NovedadAdicion = novedad
+
+	NovedadAdicionPost := make(map[string]interface{})
+	contratoid, _ := strconv.ParseInt(NovedadAdicion["contrato"].(string), 10, 32)
+	numerocdpid, _ := strconv.ParseInt(NovedadAdicion["numerocdp"].(string), 10, 32)
+	numerosolicitudentero := NovedadAdicion["numerosolicitud"].(float64)
+	numerosolicitud := strconv.FormatFloat(numerosolicitudentero, 'f', -1, 64)
+	vigencia, _ := strconv.ParseInt(NovedadAdicion["vigencia"].(string), 10, 32)
+	vigenciacdp, _ := strconv.ParseInt(NovedadAdicion["vigencia"].(string), 10, 32)
+
+	fmt.Println(NovedadAdicion["contrato"], NovedadAdicion["numerocdp"], NovedadAdicion["numerosolicitud"], NovedadAdicion["vigencia"], NovedadAdicion["vigencia"])
+	fmt.Println("\n", contratoid, numerocdpid, numerosolicitud, vigencia, vigenciacdp, "\n")
+
+	NovedadAdicionPost["NovedadPoscontractual"] = map[string]interface{}{
+		"Aclaracion":        nil,
+		"Activo":            true,
+		"ContratoId":        contratoid,
+		"FechaCreacion":     nil,
+		"FechaModificacion": nil,
+		"Id":                0,
+		"Motivo":            NovedadAdicion["motivo"],
+		"NumeroCdpId":       numerocdpid,
+		"NumeroSolicitud":   numerosolicitud,
+		"Observacion":       nil,
+		"TipoNovedad":       6,
+		"Vigencia":          vigencia,
+		"VigenciaCdp":       vigenciacdp,
+	}
+
+	fechas := make([]map[string]interface{}, 0)
+
+	fechas = append(fechas, map[string]interface{}{
+		"Activo":            true,
+		"Fecha":             NovedadAdicion["fechasolicitud"],
+		"FechaCreacion":     nil,
+		"FechaModificacion": nil,
+		"Id":                0,
+		"IdNovedadesPoscontractuales": map[string]interface{}{
+			"Id": nil,
+		},
+		"IdTipoFecha": map[string]interface{}{
+			"Id": 7,
+		},
+	})
+	fechas = append(fechas, map[string]interface{}{
+		"Activo":            true,
+		"Fecha":             NovedadAdicion["fechaadicion"],
+		"FechaCreacion":     nil,
+		"FechaModificacion": nil,
+		"Id":                0,
+		"IdNovedadesPoscontractuales": map[string]interface{}{
+			"Id": nil,
+		},
+		"IdTipoFecha": map[string]interface{}{
+			"Id": 1,
+		},
+	})
+
+	NovedadAdicionPost["Fechas"] = fechas
+
+	propiedades := make([]map[string]interface{}, 0)
+	propiedades = append(propiedades, map[string]interface{}{
+		"Activo":            true,
+		"FechaCreacion":     nil,
+		"FechaModificacion": nil,
+		"Id":                0,
+		"IdNovedadesPoscontractuales": map[string]interface{}{
+			"Id": nil,
+		},
+		"IdTipoPropiedad": map[string]interface{}{
+			"Id": 6,
+		},
+		"propiedad": NovedadAdicion["valoradicion"],
+	})
+
+	propiedades = append(propiedades, map[string]interface{}{
+		"Activo":            true,
+		"FechaCreacion":     nil,
+		"FechaModificacion": nil,
+		"Id":                0,
+		"IdNovedadesPoscontractuales": map[string]interface{}{
+			"Id": nil,
+		},
+		"IdTipoPropiedad": map[string]interface{}{
+			"Id": 2,
+		},
+		"propiedad": NovedadAdicion["cesionario"],
+	})
+
+	NovedadAdicionPost["Propiedad"] = propiedades
+
+	fmt.Println(NovedadAdicionPost)
+
+	return NovedadAdicionPost
+}
