+++ models/suspension.go
@@ -1,0 +1,126 @@
+package models
+
+import (
+	"fmt"
+	"strconv"
+)
+
+func ConstruirNovedadSuspension(novedad map[string]interface{}) (novedadformatted map[string]interface{}) {
+	NovedadSuspension := make(map[string]interface{})
+	NovedadSuspension = novedad
+
+	NovedadSuspensionPost := make(map[string]interface{})
+	contratoid, _ := strconv.ParseInt(NovedadSuspension["contrato"].(string), 10, 32)
+	//numerocdpid, _ := strconv.ParseInt(NovedadSuspension["numerocdp"].(string), 10, 32)
+	numerosolicitudentero := NovedadSuspension["numerosolicitud"].(float64)
+	numerosolicitud := strconv.FormatFloat(numerosolicitudentero, 'f', -1, 64)
+	vigencia, _ := strconv.ParseInt(NovedadSuspension["vigencia"].(string), 10, 32)
+	vigenciacdp, _ := strconv.ParseInt(NovedadSuspension["vigencia"].(string), 10, 32)
+
+	NovedadSuspensionPost["NovedadPoscontractual"] = map[string]interface{}{
+		"Aclaracion":        nil,
+		"Activo":            true,
+		"ContratoId":        contratoid,
+		"FechaCreacion":     nil,
+		"FechaModificacion": nil,
+		"Id":                0,
+		"Motivo":            NovedadSuspension["motivo"],
+		"NumeroCdpId":       0,
+		"NumeroSolicitud":   numerosolicitud,
+		"Observacion":       nil,
+		"TipoNovedad":       1,
+		"Vigencia":          vigencia,
+		"VigenciaCdp":       vigenciacdp,
+	}
+
+	fechas := make([]map[string]interface{}, 0)
+
+	fechas = append(fechas, map[string]interface{}{
+		"Activo":            true,
+		"Fecha":             NovedadSuspension["fecharegistro"],
+		"FechaCreacion":     nil,
+		"FechaModificacion": nil,
+		"Id":                0,
+		"IdNovedadesPoscontractuales": map[string]interface{}{
+			"Id": nil,
+		},
+		"IdTipoFecha": map[string]interface{}{
+			"Id": 5,
+		},
+	})
+	fechas = append(fechas, map[string]interface{}{
+		"Activo":            true,
+		"Fecha":             NovedadSuspension["fechasolicitud"],
+		"FechaCreacion":     nil,
+		"FechaModificacion": nil,
+		"Id":                0,
+		"IdNovedadesPoscontractuales": map[string]interface{}{
+			"Id": nil,
+		},
+		"IdTipoFecha": map[string]interface{}{
+			"Id": 7,
+		},
+	})
+	fechas = append(fechas, map[string]interface{}{
+		"Activo":            true,
+		"Fecha":             NovedadSuspension["fechasuspension"],
+		"FechaCreacion":     nil,
+		"FechaModificacion": nil,
+		"Id":                0,
+		"IdNovedadesPoscontractuales": map[string]interface{}{
+			"Id": nil,
+		},
+		"IdTipoFecha": map[string]interface{}{
+			"Id": 8,
+		},
+	})
+	fechas = append(fechas, map[string]interface{}{
+		"Activo":            true,
+		"Fecha":             NovedadSuspension["fechareinicio"],
+		"FechaCreacion":     nil,
+		"FechaModificacion": nil,
+		"Id":                0,
+		"IdNovedadesPoscontractuales": map[string]interface{}{
+			"Id": nil,
+		},
+		"IdTipoFecha": map[string]interface{}{
+			"Id": 6,
+		},
+	})
+
+	NovedadSuspensionPost["Fechas"] = fechas
+
+	propiedades := make([]map[string]interface{}, 0)
+	propiedades = append(propiedades, map[string]interface{}{
+		"Activo":            true,
+		"FechaCreacion":     nil,
+		"FechaModificacion": nil,
+		"Id":                0,
+		"IdNovedadesPoscontractuales": map[string]interface{}{
+			"Id": nil,
+		},
+		"IdTipoPropiedad": map[string]interface{}{
+			"Id": 3,
+		},
+		"propiedad": NovedadSuspension["periodosuspension"],
+	})
+	propiedades = append(propiedades, map[string]interface{}{
+		"Activo":            true,
+		"FechaCreacion":     nil,
+		"FechaModificacion": nil,
+		"Id":                0,
+		"IdNovedadesPoscontractuales": map[string]interface{}{
+			"Id": nil,
+		},
+		"IdTipoPropiedad": map[string]interface{}{
+			"Id": 2,
+		},
+		"propiedad": NovedadSuspension["cesionario"],
+	})
+
+	NovedadSuspensionPost["Propiedad"] = propiedades
+
+	fmt.Println(NovedadSuspensionPost)
+
+	return NovedadSuspensionPost
+}
