+++ models/terminacionanticipada.go
@@ -1,0 +1,139 @@
+package models
+
+import (
+	"fmt"
+	"strconv"
+)
+
+func ConstruirNovedadTAnticipada(novedad map[string]interface{}) (novedadformatted map[string]interface{}) {
+	NovedadTAnticipada := make(map[string]interface{})
+	NovedadTAnticipada = novedad
+
+	NovedadTAnticipadaPost := make(map[string]interface{})
+	contratoid, _ := strconv.ParseInt(NovedadTAnticipada["contrato"].(string), 10, 32)
+	//numerocdpid, _ := strconv.ParseInt(NovedadSuspension["numerocdp"].(string), 10, 32)
+	numerosolicitudentero := NovedadTAnticipada["numerosolicitud"].(float64)
+	numerosolicitud := strconv.FormatFloat(numerosolicitudentero, 'f', -1, 64)
+	vigencia, _ := strconv.ParseInt(NovedadTAnticipada["vigencia"].(string), 10, 32)
+	vigenciacdp, _ := strconv.ParseInt(NovedadTAnticipada["vigencia"].(string), 10, 32)
+
+	NovedadTAnticipadaPost["NovedadPoscontractual"] = map[string]interface{}{
+		"Aclaracion":        nil,
+		"Activo":            true,
+		"ContratoId":        contratoid,
+		"FechaCreacion":     nil,
+		"FechaModificacion": nil,
+		"Id":                0,
+		"Motivo":            NovedadTAnticipada["motivo"],
+		"NumeroCdpId":       0,
+		"NumeroSolicitud":   numerosolicitud,
+		"Observacion":       nil,
+		"TipoNovedad":       5,
+		"Vigencia":          vigencia,
+		"VigenciaCdp":       vigenciacdp,
+	}
+
+	fechas := make([]map[string]interface{}, 0)
+
+	fechas = append(fechas, map[string]interface{}{
+		"Activo":            true,
+		"Fecha":             NovedadTAnticipada["fecharegistro"],
+		"FechaCreacion":     nil,
+		"FechaModificacion": nil,
+		"Id":                0,
+		"IdNovedadesPoscontractuales": map[string]interface{}{
+			"Id": nil,
+		},
+		"IdTipoFecha": map[string]interface{}{
+			"Id": 5,
+		},
+	})
+	fechas = append(fechas, map[string]interface{}{
+		"Activo":            true,
+		"Fecha":             NovedadTAnticipada["fechasolicitud"],
+		"FechaCreacion":     nil,
+		"FechaModificacion": nil,
+		"Id":                0,
+		"IdNovedadesPoscontractuales": map[string]interface{}{
+			"Id": nil,
+		},
+		"IdTipoFecha": map[string]interface{}{
+			"Id": 7,
+		},
+	})
+	fechas = append(fechas, map[string]interface{}{
+		"Activo":            true,
+		"Fecha":             NovedadTAnticipada["fecha_terminacion_anticipada"],
+		"FechaCreacion":     nil,
+		"FechaModificacion": nil,
+		"Id":                0,
+		"IdNovedadesPoscontractuales": map[string]interface{}{
+			"Id": nil,
+		},
+		"IdTipoFecha": map[string]interface{}{
+			"Id": 9,
+		},
+	})
+
+	NovedadTAnticipadaPost["Fechas"] = fechas
+
+	propiedades := make([]map[string]interface{}, 0)
+	propiedades = append(propiedades, map[string]interface{}{
+		"Activo":            true,
+		"FechaCreacion":     nil,
+		"FechaModificacion": nil,
+		"Id":                0,
+		"IdNovedadesPoscontractuales": map[string]interface{}{
+			"Id": nil,
+		},
+		"IdTipoPropiedad": map[string]interface{}{
+			"Id": 9,
+		},
+		"propiedad": NovedadTAnticipada["numerooficioestadocuentas"],
+	})
+	propiedades = append(propiedades, map[string]interface{}{
+		"Activo":            true,
+		"FechaCreacion":     nil,
+		"FechaModificacion": nil,
+		"Id":                0,
+		"IdNovedadesPoscontractuales": map[string]interface{}{
+			"Id": nil,
+		},
+		"IdTipoPropiedad": map[string]interface{}{
+			"Id": 10,
+		},
+		"propiedad": NovedadTAnticipada["valor_desembolsado"],
+	})
+	propiedades = append(propiedades, map[string]interface{}{
+		"Activo":            true,
+		"FechaCreacion":     nil,
+		"FechaModificacion": nil,
+		"Id":                0,
+		"IdNovedadesPoscontractuales": map[string]interface{}{
+			"Id": nil,
+		},
+		"IdTipoPropiedad": map[string]interface{}{
+			"Id": 11,
+		},
+		"propiedad": NovedadTAnticipada["saldo_contratista"],
+	})
+	propiedades = append(propiedades, map[string]interface{}{
+		"Activo":            true,
+		"FechaCreacion":     nil,
+		"FechaModificacion": nil,
+		"Id":                0,
+		"IdNovedadesPoscontractuales": map[string]interface{}{
+			"Id": nil,
+		},
+		"IdTipoPropiedad": map[string]interface{}{
+			"Id": 12,
+		},
+		"propiedad": NovedadTAnticipada["saldo_universidad"],
+	})
+
+	NovedadTAnticipadaPost["Propiedad"] = propiedades
+
+	fmt.Println(NovedadTAnticipadaPost)
+
+	return NovedadTAnticipadaPost
+}
